Make eachindex more efficient for linear arrays
This greatly improves the ability to use `eachindex` in generic code.  For arrays with fast linear storage, `eachindex` now simply returns a `UnitRange` to linearly index the array. For arrays with slow linear access, this still returns a `CartesianRange`. In typical code constructs, the resulting elements from the iterators can be used identically.

As a result, we can simplify AbstractArray cartesian iteration by moving all the complexity into the `eachindex` iterator.