improve the quality of performance-tips.rst
Current performance tips (TIPS) is not very pedagogical. So I rewrite a part of TIPS to make it more understandable.
The change includes rewriting the motivation (at the beginning), adding section "why Julia can be fast", rewrite section "avoid global variables" as "be cautious to the use of global variables, especially in a loop".
Numerous examples and intuitions are provided.