{leading/count}_{zeros/ones} do not widen their arguments
These intrinsics now return a signed integer of the same width as
their argument. This contributes to type-stable code when working
with non-native integer sizes. Note that they still returned a
signed value even when their argument is unsigned. This is so that
counting zeros, for instance, does not push an entire calculation
to be unsigned.