Stopgap for Vararg widening
This is a not-very-good fix for the following problem. It turns out that
since Vararg is a type (which I would argue is a bad idea), it can get
widened into a `<:Vararg` type var. This is not handled correctly and
can cause inference to crash. This patch does not fix the fact that we
could potentially encouter something like `Tuple{<:Any}` which could
come from `Tuple{Vararg{T}}`.