Fix Profiler segfault on Darwin
We frequently find ourselves in the presence of invalid unwind information,
which can cause the profiler to try to access invalid memory. It is possible
to probe before whether the memory is accessible, but the easier solution is
just to catch the SEGV and recover. This used to work originally, but in
12d363a1c678fabec61b215b8ecc5e6b7a17cc65, I switched exception handling
to be thread-local to be able to debug the exception handler. Unfortunately,
since the profiler runs on a different thread, it was no longer covered by
this protection. Rectify that.