Throw a parser error to disambiguate numeric literals followed by dot operators
This *breaking* change throws a parser error asking the user to
disambiguate the parsing of a numeric literal and dot operator with no
whitespace in between.  We are consistent in how this is handled (always
favor the operator in parsing with no disambiguating whitespace), but
there was popular support that this should be special cased as it it
quite hard to track these subtle syntax errors down.

Closes #10920