remove redundant list of local var names from AST, add list of static params
part of #10403

ast.args[2][1] has been removed since all of its information was also
in the varinfo lists ast.args[2][2]. the varinfo lists are now at
ast.args[2][1]. args[2][2] is captured var info, args[2][3] is gensym info,
and now args[2][4] is a list of static parameter names.

GlobalRef is now built in so it can be used more extensively.