<<, >>, >>>: use intrinsics for more mixed-type shift operations.
LLVM isn't able to eliminate the additional checks done by fallbacks
for the mixed-type bitshift operations, so this change just calls
the intrinsics directly, which minimizes the amount of work done.
The fallback now avoids conversion checking as well and just returns
zero if the second argument is too large to fit in an Int.