new doc city: script to generate new helpdb.jl (md) from old (rst)
Current doc generation process generates a ReStructured text format
helpdb.jl file that is loaded when building Julia. The new doc system
will grab docstrings while building Julia and use those to generate
parts of the manual -- currently in ReStructured text, but in the
future in Markdown format.

As a step towards that, this script generates the new Markdown index
from the old ReStructured text index. You can run it as:

    julia doc/newdoc.jl
    julia doc/markdown.jl
    mv base/docs/helpdb{2,}.jl
    julia doc/coalesce.jl