new doc city: script to gen manual/stdlib docs from new helpdb.jl
The next piece of the doc up date is a script that uses help info
picked up from base/docs/helpdb.jl while building julia, to make a
new standard library. Currently, we merge the docs for all methods
into a single blob which gets included at every point where *any*
method of the function was documented previously. This is a bit
confusing but isn't the end of the world while we migrate to a new
way of generating our manual. The script can be run as:

    julia doc/genstdlib.jl

From this, you can do

	make -C doc helpdb.jl

This regenerates the old help index from the new generated manual,
allowing you to compare them. The diff is essentially all additions
so we didn't delete any essential documentation.