dlopen should follow platform default behavior
For OSX the default behavior is to dlopen with RTLD_GLOBAL while
on most other POSIX platforms the default is RTLD_LOCAL.  This reverts
the behavior introduced in #11352 which made RTLD_LOCAL the default for
all platforms which caused issues when loading shared libraries on OSX.

This PR also cleans up some `Libdl` behavior and adds tests for the
`Libdl` module.  `Libdl.dlopen` now prints a detailed error message in
the thrown exception rather than printing to STDERR.  `Libdl.dlsym` now
throws an exception instead of printing an error message to STDERR if
the symbol is not found in the shared library.  `RTLD_*` enum values now
start at 1 instead of 0 to reflect the fact that RTLD_LOCAL is not
universal default behavior on all POSIX platforms and needs to be detected on
OSX.

cc: @vtjnash, @poulson

closes #11692
fixes #11689