reimplement require/include to better support precompilation
the codepaths through these methods were not particularly amenable to adding incremental precompilation.
the problem is that the cache-load code paths need to be mirrors of their non-cache counterparts.
this minimizes the code paths to running "include" to one, and similarly for "reload"/"require"
those are now limited to only handling module names (symbols), rather than arbitrary paths

this also fixes a number of issues with the incremental serializer that
have snuck in over time.