Fix a libuv assertion after redirect_*(::IOStream)
When passing an IOStream to redirect_*, there was still a reference
to the libuv side stream, which would be used to print errors, etc. However,
since the julia side is now unreferenced, it would be gc'ed and its finalizer
would take the libuv side with it, leaving the C side pointer pointing to free'd
memory. Fix this by recording the IOStream as the new JL_STD* stream on the C
side as well.