prevent accidental shutdown->close pattern on exit
since julia finalizers were now calling uv_shutdown on all streams (including stdio), and the code to check the julia isopen status got deleted, this code now needs to ensure that the jl_close_uv function is a no-op for streams in shutdown. this was previously observed in gdb to result in a call to uv_close on all streams in the jl_atexit_hook libuv-walk.