fix poll_fd implementation (close #4401)
this eliminates the callback interface to the filesystem poll functions,
to be consistent with the changes to the rest of base

additionally, this exposes the full libuv API for watch_file and
poll_file, rather than returning inconsistent return values from these two functions

this fixes FDWatcher to ensure that it is unique for any given fd,
(to avoid a libuv assertion). it must be manually closed before closing
fd, or the applications behavior is undefined.