add Base.resolve(::GlobalRef; force=false)
this converts a GlobalRef to one that points to the module that owns
the binding. if force==true, it acts as if code in the module
actually tried to read the variable. if force==false, it makes no
state changes but returns the argument unmodified if the binding
is not fully resolved yet.