Improve build with sanitizer enabled.
Do not set LD_LIBRARY_PATH on every configure command, because some
of them use awk, which uses gmp, which can be a problem, but do set
of them use awk, which uses gmp, which can be a problem, but do set
it on PATCHELF directly if we're building it. Also special case
FORTRAN LDFLAGS for dependencies, which use gfortran for linking
(in particular ARPACK) and do not include sanitizer flags there.

Also add a patch that makes libunwind sanitizer compatible.