[OSX] Register Julia frames with both our and the system unwinder
Two years ago, we added a custom unwinder on OS X that did a better job
at unwinding from non-callsite locations by doing a small amount of assembly
profiling. Unfortunately, this also means that we were no longer registering
our JIT frames with the system unwinder. This is problematic when interfacing
with external libraries that contain calls to the unwinder, because they'll
choke at the julia frames. Fix this by registering julia frames with both
unwinders. This isn't ideal and perhaps we can find a more intelligent
way to do unwinding for julia frames, because we do have the compiler around,
but for now, this'll have to do.

This supersedes #12006, which did the same thing, but by renaming symbols,
instead of using dlsym with RTLD_NEXT.