Fixed issue with non-convertable periods.
The coarser period function allows for conversions of Milliseconds
to Seconds, Seconds to Minutes, ... and Days to Weeks. Weeks are
not convertable to Months but Months are convertable to Years. This
gap in the allowed conversions was previously not handled by the
canonicalization which attempts to have all of Period to be the
same sign.