Unify and shorten function/type docs display.
This changes the behaviour of documentation output for functions and types.
Previously all documentation was concatenated resulting in overly long
results, see `?convert` for example.

With this change only the main docs are displayed and any additional docs
are summarised in a numbered list afterwards. To view docs from this list
we can rerun the query and append a comma followed by the number:

    julia> foo
    ...

      1. foo(::Any)
      2. foo(::Int)

    julia> foo, 1
    ...