Remove specialized sort! method for PartialQuickSort{Int} (fixes #12833)
There were two sort methods for PartialQuickSort, depending
on whether it was parameterized by an Int (representing the
last index needing to be sorted) or a Range.  The version
parameterized by an Int has one less comparison per branch,
but it causes inference to fail to infer the return type of
sort/sort! under many circumstances.  Removing this version
and only providing the generic function restores proper type
inference.

If there's an underlying bug that is fixed, this can be reverted
(though the test should remain.)