Breaking change, but minor(?) for countlines()
I'm not sure this is the right place to put this. Didn't notice a section on breaking changes. Linked issue, has different title, but its second commit is on this, to docs. Should you not link issues, not that directly? I didn't find another correct issue number. Countlines has parameters, e.g. for the new, for empty lines. Should all (or any) be listed?