Fix newly added backtrace test
For any given frame the PC will be after the instruction that actually executed. In some cases, this PC can be past the end of the function boundary (or at least be not covered by line information). Thus, one should always pass the address-1 to jl_lookup_code_address (see show_backtrace in base for example).