Properly set target features for the execution engine
LLVM expects the user to pass in the target features to use manually as those
target features available for the current process may not necessarily be the
same as those available for the CPU family (e.g. due to virtualization or
vendor differences). LLVM provides the getHostCPUFeatures function to query
cpuid (on x86) or other cpu specific mechanisms for the set of supported
target architecture features. We use that function, but also keep disabling
those features that we disabled before due to LLVM bugs.

Conflicts:
	src/codegen.cpp