fix mysterious compiler performance regression
11009e152383de6f5ce665d537200b2a43f36a07 caused a large compiler
performance regression for some code. By isolating each piece of the
change, @vtjnash and I determined that this seemingly meaningless
movement of declarations in boot.jl was the cause. After poking at it
for a couple hours we still have no idea why. It's possibly related
to the order of methods in Core.call.