Minor docsystem refactoring
Combines the `doc(::Function)` and `doc(::DataType)` methods since they
were nearly identical.

Enables searching for docstrings attached to specific constructors in the
same manner as is currently allowed for normal methods. Prior to this it
resulted in a no method error.

Fixes #12701 where splatted method docs were shadowing others. Although not
the original cause of the issue, with the change to storing signatures
as `Tuple{...}` (in #12835) this meant that `f(x...)` was stored as
`Tuple{Vararg{Any}}` which is equal to `Tuple`. Since the default value of
`sig` was `Tuple` not all matching docs were returned in the example
provided in #12701.

(cherry picked from commit 1b1e6c612efbf39949f63d4d56ae9b467ff8d703)
ref #13330