Profile JIT events with OProfile
This assumes that OProfile is installed in /usr.  Adjust deps/Makefile
if it is installed elsewhere (the line '--with-oprofile=/usr').

In the current default version of LLVM (3.3), the check to see if
the execution is being profiled only looks for the older style of
running OProfile. (It checks if the 'oprofiled' daemon is running.)
The check is in
llvm-3.3/lib/ExecutionEngine/OProfileJIT/OProfileWrapper.cpp

This is fixed to recognize 'operf' in newer versions of LLVM.