Fix regression in `atrepl`
Docs attached to numbers weren't being found.

Finding docs for specific methods using the syntax

    help?> f(::Int)

wasn't being recognised due to changes in how `gen_call_with_extracted_types`
generated an error when it failed. (Previously the error was thrown
by the function, but is now returned as an `:(error("..."))` expression.)