Update noteworthy-differences.rst
I read on the users group

* https://groups.google.com/d/msg/julia-users/hmoUivSoYdE/r7ytyqa-AgAJ

that Julia uses geometric expansion, which happened to be relevant to some code I'm working on (first Julia project yay!). Coincidentally, I had just read this note about differences with Matlab and thought a link would be helpful/interesting.

Unfortunately, I am not too familiar with the markdown used on this page and could not get the link with tag to work properly. Any idea? If it is worth making this change, I can revise the link once I know how to do it.