relax Int to Integer in similar and getindex for ranges (fix #13738 )
This is a minimal fix to the error triggered by:
> r = UInt(1):UInt(2); r[r];

r[r] should produce a range (or at least an array).

The error first comes from the fact that similar(h, size(h)) selects
similar(a::AbstractArray, T) because T (meant to be a Type) is
unconstrained, and size(h) doesn't match the dims::Dims argument of
other methods.

This patch therefore relaxes Dims to DimsInteger for similar, and
constrains T. Also, r[r] should probably be a UnitRange equal to r,
like for UnitRange{Int}; the getindex methods for ranges are updated
accordingly.