Vectorize matlab parseint to improve performance
But I realize that all other parseint tests are written using loops, so I am not sure this edit is appropriate. But if not then I think you should add comments in the code to explain why it is written as it is, because it really stands out in the performance comparison table.