Ignore 'file exists' exception in mkpath
Although mkpath checks to see which parent directories need to be
created, there is a race condition from the time of that check and the
actual creation of directories. In the event another process creates one
of those directories during that time, this commit will suppress the
mkdir error due to the file already existing. Addresses issue #13094.

I put the type check before the check for new directory existence since
the type check is faster.