Rewrite JIT based on ORC
This is somewhat of a rewrite of the way we use MCJIT + a move to ORC JIT to avoid a) some runtime overhead related to MCJIT not reusing the pass manager and b) the memory overhead of MCJIT requiring you to hold onto the LLVM IR, both of which will not be fixed upstream, because they are fixed in the C API. The new ORC based jit is in jitlayers.cpp, but the intention is that both the old jit and MCJIT continue working for now. As a result the code is slightly more messy than it should be in a number places because there are now three possible JITs that could run over it.

There are two core aspects of this rewrite:
1) Attempting to combine multiple functions into a single module in order to reduce per-module overhead in the compiler. The primary difficulty here is that sometimes we need to compile functions while we're in the middle of IRgening another function (most notably in inference and in staged functions). As a result, we cannot simply have just one module into which everything gets emitted, because that way LLVM would attempt to generate code for a partially IRgened function, which is not what we want. Instead what this IR does is to temporarily put all functions into a global module (technically they could be parentless, but some of LLVM's APIs don't like dealing with parentless functions), which then get reparented to a single active module as soon as all function in a given call cycle are complete.
2) Being more aggressive about freeing memory at all stages of using LLVM. This includes freeing the IR module, as soon as the native code is generated and freeing the native code as soon as it's moved to the executable pages (both of which we were holding on to). The one caveat to this is that there is an additional copy of the native code for the debugger (for the debug sections, in the future I will make sure that copy does not include the actual generated code). Our backtracing code now shares that copy for reading DWARF information.