Fix handling of trailing spaces in REPL shell mode
As Base.shell_parse throws away trailing spaces (unless they
are escaped) we need to special case that for completions
in REPL shell mode.

Prior to this any trailing spaces were ignored for completions,
which led to problems as we tried to complete the word before
the space.

For example in case of `"ls /tmp "` (note the trailing space)
hitting tab would have made us try to complete `tmp`. Also
because the offsets are wrong in that case, the input value
would have changed to `"ls /ttmp/"` (note the double tt, the
trailing space is gone).

This also adds a regression test for completing paths with spaces
in them as I was afraid to break that.