Implement unchecked non-wrapping integer arithmetic
For each basic integer arithmetic operation (abs, negation, +, -, *, div, rem, fld, mod, cld), introduce a respective unsafe_nowrap_* function that performs no overflow checking, and that signals to LLVM that no overflow can occur via the respective NSW and NUW flags. When these functions are used, it is the responsibility of the caller to ensure that these properties hold. For correct code (i.e. where the respective checked_* functions do not report an error), the results are identical to these checked_* functions.

- Introduce intrinsics for respective LLVM intrinsics
- Add new file `base/nowrap.jl`
- Add documentation
- Add test cases