Treat macros the same as functions in docsystem
Macros are now treated as generic functions so we should naturally
handle them in the same way as functions in the docsystem.

Since the expression `@time` is no different to `@time()` this is
slightly problematic for querying the docsystem for the docs associated
with the `Function` rather than any one particular `Method` of the
macro. Hence I've added the syntax `help?> :@time` to get docs for the
whole macro while `help?> @time`, `help?> @time(x)`, etc. will get docs
for a particular signature. The quote (`:`) syntax mirrors that of

    "..."
    :@time

syntax which is already available for use to document `@time` so using
it for retrieving as well as setting docs doesn't seem too much of a
stretch.