Safer serialization of SubArrays{T,N,A<:Array}
When dealing with SubArrays-of-Arrays, we currently make serialization more efficient by "trimming" any unused data. However, we were not careful to reconstruct the exact type, and when the type was listed as a type-parameter in another type this can lead to problems. This reconstructs the type exactly, making the trimming strategy safe.

Note this only works for SubArrays-of-Arrays; we still serialize the entire parent array for any other kind of SubArray.