Revert "Disable check targets in deps during a regular build."
This reverts commit d93b819bb7ea159efbbebf271049d1c4db5b1fd6.

There are too many changes between master and release-0.4 that were
not resolved properly here.

For example, get-fftw-single is broken (and hence get-fftw, and
full-source-dist) because it refers to the srccache location that
only exists with out-of-tree-builds on master.

The dependency check targets also need to be running (and passing)
in a different target that runs when the buildbots produce releases
before this can be backported.