WIP: Try defaulting to views
This bootstraps but is barely functional due to all the SubArrays appearing in unexpected places. This will allow incremental fixes without worrying about the impossibility that is getting slices through bootstrap.  It won't tell the whole story due to # 265, but it can be a starting point that anyone can start chasing down all the `copy`s that may be required.