Added value of sinc and cosc at Infinities.
Defined the value of the sinc and cosc functions for their limiting values at plus or minus infinities.

Update values of sinc and cosc for Inf

Changed the definition only for real numbers.

added values of sinc and cosc at Inf in Real case

Added test for sinc(Inf) and cosc(Inf) ==0