Provide script `julia-native`
Provide two scripts `julia-native-setup` and `julia-native`. These are installed into the `bin` directory, intended to be called by the end user and/or administrator who installs Julia.

From the documentation:

The script 'julia-native-setup' creates an optimized version of Julia's system image for a particular CPU type. Together with the script 'julia-native', this makes it quite easy to maintain a single Julia installation on a heterogeneous cluster that has a common file system.

1. Install Julia in such a way that it works on every machine of the cluster. This usually requires either building Julia on a machine with the oldest CPU type, or explicitly setting "JULIA_CPU_TARGET".

2. For each CPU type, run 'julia-native-setup' once, on a machine with this CPU type. This creates an optimized version of Julia's system image for this CPU type. (Note: This script cannot run in parallel.)

3. To run Julia, use the script 'julia-native' instead of plain 'julia'. This script detects the machine's CPU type, and then executes Julia with the optimized system image.