julia-mode: Fix prompt regexp
Fixes #14002

The primary problem was the lack of a space at the end of the regexp,
which didn't match the output of the shell process (does have a space).
This meant that emacs sometimes placed the point one character before
the end of the buffer, which prevented M-p from working.

This new regexp also allows for other prompts, such as "shell> "