julia-mode: Improve paren indent perf
Replace julia-paren-indent with a version that uses the higher level
function backward-up-list to find any previous opening parentheses. This
function automatically skips balanced expressions like strings, as well
as comments.

On some test files the time to indent falls by more than 75% with this
change, even after increasing the threshold for paren lookback to 10k.
Note the threshold is removed in this version.