Remove `at-init` from docsystem.
Replaces macro init with a function instead. Appears to reduce the
amount of code generated by `at-doc` which should help shorten
bootstrapping time slightly.

Also changes the name of the `ObjectIdDict` used to store docstrings
to a gensym'd one to avoid naming clashes. `Docs.meta` has always
been available to access this const without needing to know it's
name so this shouldn't be a major breaking change.