prevent an item in the processing state from being recursively worked on
the interaction here is rather complex, but recursing on a frame already in typeinf_frame
causes the state upon return to that function to be a bit convoluted.
detecting frame.inferred at that point and aborting was also correct,
but requires a bit of code duplication at each abstract_interpret call
to check and return quickly.

making the processing state part of inworkq rather than stuck' should be a bit more robust
against future edits to this code