Shared Docstring References
This implements "shared" docstrings by reusing the same `DocStr`
objects when possible, i.e. when using the tuple syntax:

    "..."
    a, b

and with the semi-common idiom I've noticed recently:

    "..."
    a

    at-doc (at-doc a) b

Note that the second case uses an `at-ref` macro in this PR rather than
reusing the `at-doc` macro, i.e.

    at-doc (at-ref a) b

but it would be possible to automagically translate the second `at-doc`
into `at-ref` for users, if people would prefer that?