Update to atomics
- make atomic_cas! return the old value instead of a "success" flag
- support atomic floating-point types
- implement some atomic floating-point operations via atomic_cas!
- disable 128-bit atomic types on i686 where our LLVM install is currently broken

The success flag previously returned by atomic_cas! can be unambiguously determined from the old value that is now returned. An alternative implementation would return a tuple instead.

Atomic floating-point types are trivial to implement for load, store, compare-and-swap, and xchg. Do that. Other operations (add, subtract, min, max) can be implemented via compare-and-swap and a loop.

Some of these features are not supported by older LLVM versions; these require bitcasts as a work-around.

Atomic pointers and atomic booleans could also be supported, but are not part of this patch.