uniquerow is ic, and seems to handle hash collisions already
This commit alters the current groupslices function to return the vector uniquerow that was originally calculated within the existing unique function.  The values contained within uniquerow for cases where there are no hash collisions are actually equal to what I was calculating in array ic.  As @simonster pointed out in comment https://github.com/JuliaLang/julia/pull/14142#issuecomment-193089681 the previous commit was not taking into account hash collisions for the values in ic.  As uniquerow within unique was already calculating the values in ic, taking into account hash collisons, and updating its values accordingly, we can just return uniquerow from groupslices.  For continuity with the conversation in #14142, I currently have assigned ic as an alias for uniquerow, but that can certainly be removed.