fix #15659: bug in decompose(::BigFloat)
This bug stems from the fact that the bits of a BigFloat are left-
aligned in its storage while the bits of a BigInt are right-aligned in
its storage. When the precision of a BigFloat is divisible by the
BigInt limb size, there is no difference between these, but in other
cases, there's a substantial difference. This fix adjusts the exponent
returned by `decompose` to account for the difference in alignment.