Fix showing pairs of different type than declared
Fixes #15739. Again.

Pairs that are declared as a less precise type than that which they held, are being shown
verbosely.

This is caused by `show(IO, Pair)` calling `show_default(IO, Pair)` if the type parameters
of the `Pair` did not match the actual types contained within. e.g: `Pair{Any,Any}(1, 2)`.

Updated tests to cover this case, and be more robust.

Refactored the showing of brackets. (Removed code duplication.)