Rewrite confusing explanation of do...end syntax.
Current explanation says "In contrast to the :func:`map` example, here io is initialized by the result of open("outfile", "w")".
This is confusing: my first reading was that, whereas in the ``map`` example above, the construction was equivalent to 
`` map(anonymous_function, [A, B, C]) ``
in the example with ``open``, the construction was equivalent to
`` open("outfile", "w") |> anonymous_function! ``

Which is of course crazy.

I think what the author was trying to convey was that in the first example, the anonymous function does not try to modify contents of its arguments, whereas in the second one, the whole point is that the anon function be able to act on the stream (so if it were a named function, it'd bear an exclamation mark).

However, "result" in "io is initialized by the result of the <open function>" creates the impression that ``open`` returns that result, which is then passed to the anon function, whereas, of course, anon f. is executed within ``open``.