Use compact output for all types when showing arrays
This gets rid of the hard-coded list of types for which the compact
representation is used: all types should implement a compact output,
stripping all type information redundant with the array prefix.
We still keep a list of types for which no prefix is needed since
these are created by default for a given syntax.

Also improve docs and fix a reference to showlimited() -> showcompact().