allow iterables in fftshift and ifftshift
In the case of shifting multiple but not all axis of a multidimensional array now requires multiple fftshift commands (which, if I am correct, includes multiple array copys). The map-function is abstract enough to allow both iterables and scalar arguments, so I see no reason to not include this behaviour.