Fix AbstractArray interface docs about defining linearindexing()
The linearindexing{T<:YourType}(::Type{T}) form is needed so that
types with and without paremeters match. It is not actually needed
for the SquaresVector example, but this is the only LinearFast
example, so the simpler version can be misleading. When an
implementation offers only linearindexing(::Type{YourType}),
indexing a single value will work, but indexing with ranges will
fail with an intriguing "indexing not defined for" error.

[av skip]