Fix for removal of limit_output
This does something useful only for multiple versions of julia-0.5, since `limit_output(io)` was introduced (+1936) and then removed (+4305). Prior to that, you couldn't call `limit_output` on an `io` object.
For earlier julia versions, this just evaluates as `false`.

For, e.g., BenchmarkTools, it seems important to be able to support a range of julia-0.5 versions.