Better backtick behaviour
This extends the backtick syntax in the markdown parser to
generalise the "one tick = code" and "two ticks = math" rule
to "odd ticks = code" and "even ticks = math".

Doing this regains the nesting behaviour of backticks and allows
both inline code and math to contain arbitrary sequences of backticks
so long as enough backticks are used to wrap the code/math.