add thread/gc safety
Yu Yichao indicated the thread-safety issues and the skeleton of
a solution, and explained the need to keep references to BigInt
before ccall, and suggested to add a field at the end of MPZ.
He also pointed at the GC un-safety of the previous solution
due to the fact that finalizers can call arbitrary code. Scott
Jones suggested to disable the GC to address this specific problem.